Check for null filename when deleting the image files (full, medium, small, overlay) of an Image: attempting to create File with null filename is ambiguous (is the null a File or a String?), resulting in Grails error

+++ grails-app/controllers/edu/umn/laccore/tmi/ImageController.groovy
@@ -218,9 +218,11 @@ class ImageController {
     }
 	
 	private removeImageFile(imageInstanceDir, filename) {
-		def fileOriginal = new File(utilsService.getImagesDir() + "/" + imageInstanceDir, filename)
-		if ( fileOriginal.exists() )
-			fileOriginal.delete()
+		if (filename) {
+			def fileOriginal = new File(utilsService.getImagesDir() + "/" + imageInstanceDir, filename)
+			if ( fileOriginal.exists() )
+				fileOriginal.delete()
+		}
 	}
 	
 	private removeImagesFromDisk(imageInstance) {
